+++ tests/js/session-tests.js
@@ -193,6 +193,50 @@ module.exports = {
             });
     },
 
+    testDefaultRealmFromUser() {
+        if (!isNodeProccess) {
+            return;
+        }
+
+        const username = uuid();
+        const realmName = 'default';
+        const expectedObjectsCount = 3;
+
+        let user;
+        return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
+            .then(() => Realm.Sync.User.login('http://localhost:9080', username, 'password'))
+            .then(u => {
+                user = u;
+                return Realm.open(Realm.automaticSyncConfiguration(user));
+            })
+            .then(realm => {
+                let actualObjectsCount = realm.objects('Dog').length;
+                TestCase.assertEqual(actualObjectsCount, expectedObjectsCount, "Synced realm does not contain the expected objects count");
+
+                const session = realm.syncSession;
+                TestCase.assertInstanceOf(session, Realm.Sync.Session);
+                TestCase.assertEqual(session.user.identity, user.identity);
+                TestCase.assertEqual(session.state, 'active');
+            });
+    },
+
+    testDefaultRealmInvalidArguments() {
+        if (!isNodeProccess) {
+            return;
+        }
+
+        const username = uuid();
+        const realmName = 'default';
+        const expectedObjectsCount = 3;
+
+        let user;
+        return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
+            .then(() => Realm.Sync.User.login('http://localhost:9080', username, 'password'))
+            .then(u => {
+                TestCase.assertThrows(() => Realm.automaticSyncConfiguration('foo', 'bar')); // too many arguments
+            })
+    },
+
     testRealmOpenWithExistingLocalRealm() {
         if (!isNodeProccess) {
             return;